requirements.txt
Flask==2.1.2
python-dotenv==0.20.0
supabase==0.4.2
Flask-Session==0.4.0
